requirements.txt
Django==2.0.6
Unipath==1.1
dj-database-url==0.5.0
django-extensions==2.0.7
django-suit==0.2.26
python-decouple==3.1
ipython==6.4.0
django-tinymce==2.7.0
gunicorn==19.9.0
django-heroku==0.3.1
django-storages==1.6.6
whitenoise==3.3.1
boto3==1.7.61
psycopg2-binary==2.7.5
django-yamlfield==1.0.3
Pillow==5.2.0
sorl-thumbnail==12.4.1
requests==2.19.1
raven==6.9.0
django-redis==4.9.0
redis==2.10.6
